controler: moved disk and warc wq watchers to a new package and created Stop functions for both watchers
